wxauto-bridge/requirements.txt
# wxauto-bridge 依赖
# wxauto 库从本地路径导入 (my-docs/wxauto-main)
# 或从 GitHub 安装: pip install git+https://github.com/cluic/wxauto.git

websockets>=12.0
pywin32>=306
comtypes>=1.2.0
uiautomation>=2.0.18
pillow>=10.0.0
psutil>=5.9.0
pywebview>=4.0
websockets>=11.0
websocket-client>=1.0